test: split `EigenSystem` tests into `test_linalg_eigen.py`
Move the `EigenSystem` test class out of test_linalg.py into a dedicated
`test_linalg_eigen.py` and rename it from `TestLinalgEigenSystemTC` to
`TestEigenSystemTC`.

The tests themselves remain unchanged.
